Fix F1-cycle for #17131
