+++ src/renderer/src/lib/pane-manager/pane-terminal-output-drain.ts
@@ -1,122 +1,26 @@
-import { recordRendererCrashBreadcrumb } from '@/lib/crash-breadcrumb-recorder'
 import {
   recordTerminalOutputQueueDebugPressure as recordQueueDebugPressure,
   terminalOutputSchedulerDebugEnabled as debugEnabled,
   terminalOutputSchedulerDebugState as debugState
 } from './pane-terminal-output-scheduler-debug'
+import { clearForegroundRelease, isEntryDrainable } from './pane-terminal-foreground-queue-state'
+import { hasHighPriorityBacklog, hasQueuedChunks } from './pane-terminal-output-queue-backlog'
 import {
-  clearForegroundHoldSafety,
-  clearForegroundRelease,
-  isEntryDrainable
-} from './pane-terminal-foreground-queue-state'
-import {
-  ALWAYS_REFRESH_FOREGROUND_SYNCHRONOUSLY,
-  BACKGROUND_BACKLOG_WARNING,
   BACKGROUND_DRAIN_INTERVAL_MS,
   DRAIN_TIME_BUDGET_MS,
-  FOREGROUND_BACKLOG_WARNING,
   HIGH_PRIORITY_DRAIN_INTERVAL_MS,
   HIGH_PRIORITY_MAX_WRITES_PER_DRAIN,
   LARGE_BACKLOG_CHARS,
-  MAX_BACKGROUND_QUEUE_CHUNKS,
   MAX_WRITES_PER_DRAIN,
-  getTerminalOutputMaxQueueChars,
   isMessageChannelDrainEnabled,
   markTerminalOutputDrainStarted,
   queuedByTerminal,
   scheduleDrain,
-  type QueueEntry,
-  type TerminalOutputBeforeWrite
-} from './pane-terminal-output-scheduler'
+  setTerminalOutputDrainRunner,
+  type QueueEntry
+} from './pane-terminal-output-queue-registry'
 
 import { writeQueuedChunk } from './pane-terminal-output-pipeline'
-// Why: every discard path MUST fire these before clearing/replacing the queue — a dropped chunk still counts as consumed, or main's in-flight window shrinks permanently and the PTY wedges.
-export function fireQueuedAckCredits(entry: QueueEntry): void {
-  for (let index = entry.chunkIndex; index < entry.chunks.length; index += 1) {
-    entry.chunks[index].ackCredit?.()
-  }
-}
-
-export function discardDetachedQueueEntry(entry: QueueEntry): void {
-  fireQueuedAckCredits(entry)
-  entry.chunks.length = 0
-  entry.chunkIndex = 0
-  entry.queuedChars = 0
-  entry.highPriority = false
-  clearForegroundRelease(entry)
-}
-
-export function queueCapExceeded(entry: QueueEntry): boolean {
-  return (
-    entry.queuedChars > getTerminalOutputMaxQueueChars() ||
-    entry.chunks.length - entry.chunkIndex > MAX_BACKGROUND_QUEUE_CHUNKS
-  )
-}
-
-export function replaceBacklogWithWarning(
-  entry: QueueEntry,
-  warning: string = BACKGROUND_BACKLOG_WARNING
-): void {
-  const shouldNotify = !entry.backgroundBacklogDropped
-  if (shouldNotify) {
-    // Why: field visibility for cap tuning — drop frequency and size decide whether the cap is too small (issue #2836 / #7017).
-    recordRendererCrashBreadcrumb('terminal_output_backlog_dropped', {
-      foreground: warning === FOREGROUND_BACKLOG_WARNING,
-      droppedChars: entry.queuedChars,
-      capChars: getTerminalOutputMaxQueueChars()
-    })
-  }
-  let beforeWrite: TerminalOutputBeforeWrite | undefined
-  for (let index = entry.chunks.length - 1; index >= entry.chunkIndex; index--) {
-    if (entry.chunks[index]?.beforeWrite) {
-      beforeWrite = entry.chunks[index].beforeWrite
-      break
-    }
-  }
-  clearForegroundHoldSafety(entry)
-  fireQueuedAckCredits(entry)
-  entry.chunks = [
-    {
-      data: warning,
-      retainedChars: warning.length,
-      foreground: false,
-      forceForegroundRefresh: false,
-      followupForegroundRefresh: false,
-      shouldRefreshForegroundSynchronously: ALWAYS_REFRESH_FOREGROUND_SYNCHRONOUSLY,
-      stripTransientCursorShows: false,
-      beforeWrite
-    }
-  ]
-  entry.chunkIndex = 0
-  entry.queuedChars = warning.length
-  entry.backgroundBacklogDropped = true
-  entry.highPriority = true
-  entry.foregroundHold = false
-  if (debugEnabled && shouldNotify) {
-    debugState.droppedBacklogCount++
-  }
-  clearForegroundRelease(entry)
-  recordQueueDebugPressure()
-  if (shouldNotify) {
-    entry.onBackgroundBacklogDropped?.()
-  }
-}
-
-export function hasQueuedChunks(entry: QueueEntry): boolean {
-  return entry.chunkIndex < entry.chunks.length
-}
-
-export function hasHighPriorityBacklog(): boolean {
-  for (const entry of queuedByTerminal.values()) {
-    if (
-      isEntryDrainable(entry) &&
-      (entry.highPriority || entry.queuedChars > LARGE_BACKLOG_CHARS)
-    ) {
-      return true
-    }
-  }
-  return false
-}
 
 function hasDrainableBacklog(): boolean {
   for (const entry of queuedByTerminal.values()) {
@@ -219,3 +123,6 @@ export function drainQueuedOutputImpl(): void {
     )
   }
 }
+
+// Why at module scope: the registry's scheduleDrain must never fire before a runner exists, so registration happens as this module is evaluated rather than on first use.
+setTerminalOutputDrainRunner(drainQueuedOutputImpl)
